refactor(challenge): my-challenge/section 응집 클러스터를 하위폴더로 그룹핑
40+ 파일 평면 폴더에서 뚜렷한 두 클러스터만 하위폴더로 분리(기존 구조 존중):
- MissionGuide* 6개 → section/mission-guide/
- MissionSubmit* 5개 → section/mission-submit/
git rename으로 히스토리 보존. import 경로(외부 app 3곳 + 내부 상대경로) 정리.
동작 변경 없음. tsc 0 / lint 통과.
