+++ apps/web/src/app/(user)/challenge/[applicationId]/[programId]/me/page.tsx
@@ -4,9 +4,9 @@ import { AsyncBoundary } from '@/common/boundary/AsyncBoundary';
 import LoadingContainer from '@/common/loading/LoadingContainer';
 import { useCurrentChallenge } from '@/context/CurrentChallengeProvider';
 import MissionCalendarSection from '@/domain/challenge/my-challenge/section/MissionCalendarSection';
-import MissionGuideSection from '@/domain/challenge/my-challenge/section/MissionGuideSection';
+import MissionGuideSection from '@/domain/challenge/my-challenge/section/mission-guide/MissionGuideSection';
 import MissionMentorCommentSection from '@/domain/challenge/my-challenge/section/MissionMentorCommentSection';
-import MissionSubmitSection from '@/domain/challenge/my-challenge/section/MissionSubmitSection';
+import MissionSubmitSection from '@/domain/challenge/my-challenge/section/mission-submit/MissionSubmitSection';
 import { useChallengeProgram } from '@/hooks/useChallengeProgram';
 import { useExperienceLevel } from '@/hooks/useExperienceLevel';
 import { useFilteredSchedules } from '@/hooks/useFilteredSchedules';
