+++ apps/web/src/domain/challenge/my-challenge/section/MyDocUploadSection.tsx
@@ -1,6 +1,6 @@
 import LoadingContainer from '@/common/loading/LoadingContainer';
 import { clsx } from 'clsx';
-import { UploadedFiles } from './MissionSubmitTalentPoolSection';
+import { UploadedFiles } from './mission-submit/MissionSubmitTalentPoolSection';
 import { DocumentFileItem } from './my-doc-upload/DocumentFileItem';
 import { useMyDocUpload } from './my-doc-upload/useMyDocUpload';
