+++ apps/web/src/domain/challenge/my-challenge/section/mission-guide/MissionGuideSection.tsx
