+++ apps/web/src/domain/challenge/my-challenge/section/mission-guide/MissionGuideSkeleton.tsx
