+++ apps/web/src/domain/challenge/my-challenge/section/mission-guide/MissionGuideZeroSection.tsx
@@ -1,9 +1,9 @@
 import { UserChallengeMissionWithAttendance } from '@/schema';
 import { clsx } from 'clsx';
 import { Dayjs } from 'dayjs';
-import MissionFileLink from '../mission/MissionFileLink';
+import MissionFileLink from '../../mission/MissionFileLink';
 import MissionGuideSkeleton from './MissionGuideSkeleton';
-import MissionHeaderSection from './MissionHeaderSection';
+import MissionHeaderSection from '../MissionHeaderSection';
 
 interface MissionGuideZeroSectionProps {
   className?: string;
